fix: fix TaskDetailModal
