+++ tasks/components/dnd-kit/Draggable.tsx
@@ -10,6 +10,7 @@ export type DraggableProps<DraggableData, DroppableData> = {
   children: ((draggableBuilderProps: DraggableBuilderProps<DraggableData, DroppableData>) => React.ReactNode | undefined),
   id: string,
   data: DraggableData,
+  className?: string,
 }
 
 /**
@@ -18,7 +19,8 @@ export type DraggableProps<DraggableData, DroppableData> = {
 export const Draggable = <DraggableData, DroppableData>({
   children,
   id,
-  data
+  data,
+  className,
 }: DraggableProps<DraggableData, DroppableData>) => {
   const { attributes, listeners, setNodeRef, transform, ...draggableBuilderProps } = useDraggable<DraggableData, DroppableData>({
     id,
@@ -30,7 +32,7 @@ export const Draggable = <DraggableData, DroppableData>({
   }
 
   return (
-    <div ref={setNodeRef} {...style} {...listeners} {...attributes}>
+    <div ref={setNodeRef} {...style} {...listeners} {...attributes} className={className}>
       {children(draggableBuilderProps)}
     </div>
   )
