+++ tasks/package.json
@@ -15,7 +15,7 @@
     "@dnd-kit/sortable": "7.0.2",
     "@headlessui/react": "1.7.19",
     "@helpwave/api-services": "workspace:*",
-    "@helpwave/hightide": "^0.0.17",
+    "@helpwave/hightide": "^0.0.19",
     "@radix-ui/react-checkbox": "1.1.3",
     "@tailwindcss/postcss": "^4.1.3",
     "@tanstack/react-query": "^4.36.1",
